# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel Survey API
  description: |
    A comprehensive survey management API built with Laravel, featuring:
    - Multi-role authentication (Admin, User, LINE OA)
    - Real-time survey updates via WebSocket broadcasting
    - Advanced analytics and reporting
    - Export/Import functionality for surveys and responses
    - Tiered rate limiting for security
    - Email verification and password reset
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Use JWT token obtained from authentication endpoints
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        email_verified_at:
          type: string
          format: date-time
          nullable: true
          example: "2024-07-21T10:30:00Z"
        role:
          type: string
          enum: [admin, user]
          example: "user"
        created_at:
          type: string
          format: date-time
          example: "2024-07-21T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-07-21T10:30:00Z"
    LineOAUser:
      type: object
      properties:
        id:
          type: integer
          example: 1
        line_user_id:
          type: string
          example: "U1234567890abcdef"
        display_name:
          type: string
          example: "John LINE User"
        role:
          type: string
          enum: [admin, user]
          example: "user"
        created_at:
          type: string
          format: date-time
          example: "2024-07-21T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-07-21T10:30:00Z"
    Survey:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Customer Satisfaction Survey"
        description:
          type: string
          nullable: true
          example: "Please rate your experience with our service"
        status:
          type: string
          enum: [draft, active, closed]
          example: "active"
        created_at:
          type: string
          format: date-time
          example: "2024-07-21T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-07-21T10:30:00Z"
        questions:
          type: array
          items:
            $ref: '#/components/schemas/SurveyQuestion'
    SurveyQuestion:
      type: object
      properties:
        id:
          type: integer
          example: 1
        survey_id:
          type: integer
          example: 1
        question:
          type: string
          example: "How satisfied are you with our service?"
        type:
          type: string
          enum: [text, radio, checkbox, select, textarea, number]
          example: "radio"
        options:
          type: string
          nullable: true
          example: "Very Satisfied,Satisfied,Neutral,Unsatisfied,Very Unsatisfied"
        required:
          type: boolean
          example: true
        order:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
          example: "2024-07-21T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-07-21T10:30:00Z"
    SurveyResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        survey_id:
          type: integer
          example: 1
        user_id:
          type: integer
          nullable: true
          example: 1
        line_user_id:
          type: integer
          nullable: true
          example: 1
        response_data:
          type: object
          example: {"1": "Very Satisfied", "2": "Great service!"}
        ip_address:
          type: string
          nullable: true
          example: "192.168.1.1"
        user_agent:
          type: string
          nullable: true
          example: "Mozilla/5.0..."
        created_at:
          type: string
          format: date-time
          example: "2024-07-21T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-07-21T10:30:00Z"
        survey:
          $ref: '#/components/schemas/Survey'
        user:
          $ref: '#/components/schemas/User'
        line_user:
          $ref: '#/components/schemas/LineOAUser'

    # Request/Response Models
...
    SurveyResponseRequest:
      type: object
      required:
        - survey_id
        - response_data
      properties:
        survey_id:
          type: integer
          example: 1
        response_data:
          type: object
          example: {"1": "Very Satisfied", "2": "Great service!"}

    # Response Wrappers
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Validation error"
        errors:
          type: object
          nullable: true
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "The given data was invalid."
        errors:
          type: object
          example:
            email: ["The email field is required."]
            password: ["The password field is required."]
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Unauthenticated"
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Resource not found"
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Too Many Requests"
...
paths:
...
  /v1/survey-responses:
    post:
      tags:
        - Survey Responses
      summary: Submit survey response
      description: Submit a response to a survey
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SurveyResponseRequest'
      responses:
        '201':
          description: Response submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Survey response submitted successfully"
                  data:
                    $ref: '#/components/schemas/SurveyResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /v1/survey-responses/{surveyResponse}:
    get:
      tags:
        - Survey Responses
      summary: Get survey response
      description: Retrieve a specific survey response
      security:
        - BearerAuth: []
      parameters:
        - name: surveyResponse
          in: path
          required: true
          schema:
            type: integer
          description: Survey Response ID
      responses:
        '200':
          description: Survey response retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/SurveyResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'

  # Admin Survey Management
...